+++ controllers/authController.js
@@ -1,221 +1,221 @@
-const User = require("../models/user");
-const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
-const {
-  generateToken,
-  generateRefreshToken,
-} = require("../utils/generateToken");
-const sendResponse = require("../utils/sendResponse");
-const sendEmail = require("../utils/sendEmail");
-const sendOTP = require("../utils/sendOTP");
+// const User = require("../models/user");
+// const bcrypt = require("bcrypt");
+// const jwt = require("jsonwebtoken");
+// const {
+//   generateToken,
+//   generateRefreshToken,
+// } = require("../utils/generateToken");
+// const sendResponse = require("../utils/sendResponse");
+// const sendEmail = require("../utils/sendEmail");
+// const sendOTP = require("../utils/sendOTP");
 
-const signup = async (req, res, next) => {
-  try {
-    const { name, email, password, role } = req.body;
-    const existingUser = await User.findOne({ email });
-    if (existingUser)
-      return sendResponse(res, 400, false, "Email already in use");
+// const signup = async (req, res, next) => {
+//   try {
+//     const { name, email, password, role } = req.body;
+//     const existingUser = await User.findOne({ email });
+//     if (existingUser)
+//       return sendResponse(res, 400, false, "Email already in use");
 
-    const hashedPassword = await bcrypt.hash(password, 10);
-    const newUser = await User.create({
-      name,
-      email,
-      password: hashedPassword,
-      role,
-      isVerified: false,
+//     const hashedPassword = await bcrypt.hash(password, 10);
+//     const newUser = await User.create({
+//       name,
+//       email,
+//       password: hashedPassword,
+//       role,
+//       isVerified: false,
      
-    });
+//     });
     
-   const otp= await sendOTP(newUser._id, newUser.email);
+//    const otp= await sendOTP(newUser._id, newUser.email);
 
-    newUser.otp = otp;
-    await newUser.save();
+//     newUser.otp = otp;
+//     await newUser.save();
 
-    const token = generateToken({
-      id: newUser._id,
-      email: newUser.email,
-      role: newUser.role,
-    });
+//     const token = generateToken({
+//       id: newUser._id,
+//       email: newUser.email,
+//       role: newUser.role,
+//     });
 
-    const refreshToken = generateRefreshToken({
-      id: newUser._id,
-      email: newUser.email,
-      role: newUser.role,
-    });
+//     const refreshToken = generateRefreshToken({
+//       id: newUser._id,
+//       email: newUser.email,
+//       role: newUser.role,
+//     });
 
-    sendResponse(res, 201, true, "User Registered Successfully", {
-      user: {
-        id: newUser._id,
-        name: newUser.name,
-        email: newUser.email,
-        role: newUser.role,
-      },
-      token,
-      refreshToken,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
+//     sendResponse(res, 201, true, "User Registered Successfully", {
+//       user: {
+//         id: newUser._id,
+//         name: newUser.name,
+//         email: newUser.email,
+//         role: newUser.role,
+//       },
+//       token,
+//       refreshToken,
+//     });
+//   } catch (error) {
+//     next(error);
+//   }
+// };
 
-const login = async (req, res, next) => {
-  try {
-    const { email, password } = req.body;
-    const user = await User.findOne({ email });
-    if (!user) return sendResponse(res, 404, false, "User Not Found");
+// const login = async (req, res, next) => {
+//   try {
+//     const { email, password } = req.body;
+//     const user = await User.findOne({ email });
+//     if (!user) return sendResponse(res, 404, false, "User Not Found");
 
-    const isMatch = await bcrypt.compare(password, user.password);
-    if (!isMatch) return sendResponse(res, 401, false, "Invalid Credentials");
+//     const isMatch = await bcrypt.compare(password, user.password);
+//     if (!isMatch) return sendResponse(res, 401, false, "Invalid Credentials");
 
-    const token = generateToken({
-      id: user._id,
-      email: user.email,
-      role: user.role,
-    });
+//     const token = generateToken({
+//       id: user._id,
+//       email: user.email,
+//       role: user.role,
+//     });
 
-    const refreshToken = generateRefreshToken({
-      id: user._id,
-      email: user.email,
-      role: user.role,
-    });
+//     const refreshToken = generateRefreshToken({
+//       id: user._id,
+//       email: user.email,
+//       role: user.role,
+//     });
 
-    sendResponse(res, 200, true, "Login Successful", {
-      token,
-      refreshToken,
-      user: {
-        id: user._id,
-        name: user.name,
-        email: user.email,
-        role: user.role,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-};
+//     sendResponse(res, 200, true, "Login Successful", {
+//       token,
+//       refreshToken,
+//       user: {
+//         id: user._id,
+//         name: user.name,
+//         email: user.email,
+//         role: user.role,
+//       },
+//     });
+//   } catch (error) {
+//     next(error);
+//   }
+// };
 
-const forgetPassword = async (req, res, next) => {
-  try {
-    const { email } = req.body;
-    const user = await User.findOne({ email });
-    if (!user) return sendResponse(res, 404, false, "User Not Found");
+// const forgetPassword = async (req, res, next) => {
+//   try {
+//     const { email } = req.body;
+//     const user = await User.findOne({ email });
+//     if (!user) return sendResponse(res, 404, false, "User Not Found");
 
-    const otp = Math.floor(1000 + Math.random() * 9000);
-    user.otp = otp;
-    await user.save();
+//     const otp = Math.floor(1000 + Math.random() * 9000);
+//     user.otp = otp;
+//     await user.save();
 
-    const html = `<h2>Password Reset OTP</h2><p>Your OTP is: <strong>${otp}</strong></p>`;
-    await sendEmail(user.email, "Password Reset OTP", html);
+//     const html = `<h2>Password Reset OTP</h2><p>Your OTP is: <strong>${otp}</strong></p>`;
+//     await sendEmail(user.email, "Password Reset OTP", html);
 
-    sendResponse(res, 200, true, "OTP Sent to Email");
-  } catch (error) {
-    next(error);
-  }
-};
+//     sendResponse(res, 200, true, "OTP Sent to Email");
+//   } catch (error) {
+//     next(error);
+//   }
+// };
 
-const resetPassword = async (req, res, next) => {
-  try {
-    const { email, OTP, password } = req.body;
-    const user = await User.findOne({ email });
-    if (!user) return sendResponse(res, 404, false, "User Not Found");
-    if (Number(OTP) !== user.otp)
-      return sendResponse(res, 400, false, "Invalid OTP");
+// const resetPassword = async (req, res, next) => {
+//   try {
+//     const { email, OTP, password } = req.body;
+//     const user = await User.findOne({ email });
+//     if (!user) return sendResponse(res, 404, false, "User Not Found");
+//     if (Number(OTP) !== user.otp)
+//       return sendResponse(res, 400, false, "Invalid OTP");
 
-    user.password = await bcrypt.hash(password, 10);
-    user.otp = null;
-    await user.save();
+//     user.password = await bcrypt.hash(password, 10);
+//     user.otp = null;
+//     await user.save();
 
-    sendResponse(res, 200, true, "Password Updated Successfully");
-  } catch (error) {
-    next(error);
-  }
-};
+//     sendResponse(res, 200, true, "Password Updated Successfully");
+//   } catch (error) {
+//     next(error);
+//   }
+// };
 
-const changePassword = async (req, res, next) => {
-  try {
-    const { currentPassword, newPassword } = req.body;
+// const changePassword = async (req, res, next) => {
+//   try {
+//     const { currentPassword, newPassword } = req.body;
 
-    if (!req.user || !req.user.id) {
-      console.log("req.user:", req.user);
-      return sendResponse(res, 401, false, "Unauthorized access");
-    }
+//     if (!req.user || !req.user.id) {
+//       console.log("req.user:", req.user);
+//       return sendResponse(res, 401, false, "Unauthorized access");
+//     }
 
-    const user = await User.findById(req.user.id).select("+password");
-    if (!user) return sendResponse(res, 404, false, "User Not Found");
+//     const user = await User.findById(req.user.id).select("+password");
+//     if (!user) return sendResponse(res, 404, false, "User Not Found");
 
-    const isMatch = await bcrypt.compare(currentPassword, user.password);
-    if (!isMatch) {
-      return sendResponse(res, 400, false, "Incorrect current password");
-    }
+//     const isMatch = await bcrypt.compare(currentPassword, user.password);
+//     if (!isMatch) {
+//       return sendResponse(res, 400, false, "Incorrect current password");
+//     }
 
-    const isSame = await bcrypt.compare(newPassword, user.password);
-    if (isSame) {
-      return sendResponse(
-        res,
-        400,
-        false,
-        "New password cannot be the same as current"
-      );
-    }
+//     const isSame = await bcrypt.compare(newPassword, user.password);
+//     if (isSame) {
+//       return sendResponse(
+//         res,
+//         400,
+//         false,
+//         "New password cannot be the same as current"
+//       );
+//     }
 
-    user.password = await bcrypt.hash(newPassword, 10);
-    await user.save();
+//     user.password = await bcrypt.hash(newPassword, 10);
+//     await user.save();
 
-    const accessToken = generateToken({
-      id: user._id,
-      email: user.email,
-      role: user.role,
-    });
+//     const accessToken = generateToken({
+//       id: user._id,
+//       email: user.email,
+//       role: user.role,
+//     });
 
-    const refreshToken = generateRefreshToken({
-      id: user._id,
-      email: user.email,
-      role: user.role,
-    });
+//     const refreshToken = generateRefreshToken({
+//       id: user._id,
+//       email: user.email,
+//       role: user.role,
+//     });
 
-    sendResponse(res, 200, true, "Password Changed Successfully", {
-      accessToken,
-      refreshToken,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
+//     sendResponse(res, 200, true, "Password Changed Successfully", {
+//       accessToken,
+//       refreshToken,
+//     });
+//   } catch (error) {
+//     next(error);
+//   }
+// };
 
-const logout = (req, res) => {
-  sendResponse(res, 200, true, "Logged out successfully");
-};
+// const logout = (req, res) => {
+//   sendResponse(res, 200, true, "Logged out successfully");
+// };
 
-const refreshAccessToken = (req, res, next) => {
-  const authHeader = req.headers["authorization"];
-  const token = authHeader?.split(" ")[1];
+// const refreshAccessToken = (req, res, next) => {
+//   const authHeader = req.headers["authorization"];
+//   const token = authHeader?.split(" ")[1];
 
-  if (!token) {
-    return sendResponse(res, 401, false, "No refresh token found");
-  }
+//   if (!token) {
+//     return sendResponse(res, 401, false, "No refresh token found");
+//   }
 
-  try {
-    const decoded = jwt.verify(token, process.env.REFRESH_SECRET);
+//   try {
+//     const decoded = jwt.verify(token, process.env.REFRESH_SECRET);
 
-    const accessToken = generateToken({
-      id: decoded.id,
-      email: decoded.email,
-      role: decoded.role,
-    });
+//     const accessToken = generateToken({
+//       id: decoded.id,
+//       email: decoded.email,
+//       role: decoded.role,
+//     });
 
-    sendResponse(res, 200, true, { accessToken });
-  } catch (error) {
-    next(error);
-  }
-};
+//     sendResponse(res, 200, true, { accessToken });
+//   } catch (error) {
+//     next(error);
+//   }
+// };
 
 
-module.exports = {
-  signup,
-  login,
-  forgetPassword,
-  resetPassword,
-  changePassword,
-  refreshAccessToken,
-  logout,
-};
+// module.exports = {
+//   signup,
+//   login,
+//   forgetPassword,
+//   resetPassword,
+//   changePassword,
+//   refreshAccessToken,
+//   logout,
+// };
